reorganized the services folder to support more changes in the future without getting lost in files:
