--- app/services/destroy_articles_service.rb
@@ -1,13 +1,0 @@
-class DestroyArticlesService
-    
-    def initialize(article_params, current_user)
-        @article = Article.find(article_params[:id])
-        @user = current_user
-    end
-
-    def call 
-        @article.uploads.purge
-        @article.destroy
-    end
-
-end
